+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Fixed a bug in the function `S_to_index_points_in_rule` that caused it to crash when the type of a feature was not specified in the solver.
 - Fixed a bug in the function `S_to_index_points_in_rule` that caused it to crash when a modality was not in the filtered DataFrame.
+- Fixed a bug in the function `S_to_index_points_in_rule` that caused it to not return the good index under some circumstances.
 
 ## 0.1.13 (2025-04-09)